requirements.txt
aiofiles
aiohttp
aioshutil
anytree
apscheduler
aria2p
asyncio
beautifulsoup4
cloudscraper
dnspython
feedparser
flask
gevent
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gTTS
gunicorn
heroku3
html-telegraph-poster
httpx
lxml
megasdkrestclient
motor
mutagen
natsort
pillow
psutil
pybase64
pymongo
pyrogram==2.0.73
python-dotenv
python-magic
qbittorrent-api
requests
speedtest-cli
telegraph
tenacity
tgcrypto
uvloop
waybackpy
xattr
yt-dlp